feat: [CI-4730]: move module
