+++ ci-manager/service/src/main/java/io/harness/ci/app/telemetry/CITelemetryStatusRepository.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.repositories;
+package io.harness.app.telemetry;
 
 import io.harness.annotation.HarnessRepo;
 import io.harness.ci.beans.entities.CITelemetrySentStatus;
